update code, harmonize policy var
Implemented most of the suggestion.

I still have the suffix _regional and _national for school_closure. We will need a discussion about school_closure as it is introducing a LOT of noise since we aggregate adm3 level policy to adm1. In addition, school_closure_national is super colinear with the lockdown, so I would like to keep those separate for now.

Sol originally kept the regional variation only. If I aggregate the 3 in "school_closure" we won't be able to separate post _processed